rename fetch suite to fetchsync
