+++ tests/integration/fetchsync/fetchsync.js
