+++ tests/integration/handlers.js
@@ -1,5 +1,5 @@
 export { handler as btoa } from './btoa/btoa.js';
 export { handler as performance } from './performance/performance.js';
 export { handler as crypto } from './crypto/crypto.js';
-export { handler as fetch } from './fetch/fetch.js';
+export { handler as fetchsync } from './fetchsync/fetchsync.js';
 export { handler as timers } from './timers/timers.js';
