+++ tests/tests.cmake
@@ -38,5 +38,5 @@ test_e2e(tla-runtime-resolve)
 test_integration(btoa)
 test_integration(performance)
 test_integration(crypto)
-test_integration(fetch)
+test_integration(fetchsync)
 test_integration(timers)
